@romanz 2024-08-19
`TCP_NODELAY` has been set on p2p connections by https://github.com/bitcoin/bitcoin/pull/6867.