@pinheadmz 2024-08-19
concept ACK, I'm kinda surprised we don't have this already. I did a quick check into libevent and don't see any helpers there like `evutil_make_listen_socket_reuseable()` (which sets `SO_REUSEADDR` so this might be the best approach